miniupnpd: update to 2.2 (20201129 git snapshot)
A few changes/fixes on top of upstream:

 - Updated new configure script with Asuswrt-specific code from old genconfig.sh
 - Fixed config.h path in ifacewatcher
 - Now use upstream Makefile.linux instead of (outdated) custom Makefile.merlin,
   with minimal changes to it:
     . Added EXTRACFLAGS content
     . Disabled building test executables
 - iptables path now passed as argument to new configure script
 - Don't append git tag to version string

This has only been tested with SDK 6.37/iptables 1.4.15 and newer (untested on
older MIPS platform since we no longer support them anyway)

+++ release/src/router/miniupnpd/testasyncsendto.c
@@ -1,8 +1,8 @@
-/* $Id: testasyncsendto.c,v 1.5 2018/03/13 10:25:52 nanard Exp $ */
+/* $Id: testasyncsendto.c,v 1.6 2019/09/24 11:48:47 nanard Exp $ */
 /* vim: tabstop=4 shiftwidth=4 noexpandtab
  * MiniUPnP project
  * http://miniupnp.free.fr/ or https://miniupnp.tuxfamily.org/
- * (c) 2006-2018 Thomas Bernard
+ * (c) 2006-2019 Thomas Bernard
  * This software is subject to the conditions detailed
  * in the LICENCE file provided within the distribution */
 
@@ -74,8 +74,8 @@ int test(void)
 		int max_fd;
 		struct timeval timeout;
 		struct timeval now;
-		syslog(LOG_DEBUG, "get_next_scheduled_send : %d next_send=%ld.%06ld",
-		       i, (long)next_send.tv_sec, (long)next_send.tv_usec);
+		syslog(LOG_DEBUG, "get_next_scheduled_send : %d next_send=%lld.%06ld",
+		       i, (long long)next_send.tv_sec, (long)next_send.tv_usec);
 		FD_ZERO(&writefds);
 		max_fd = 0;
 		gettimeofday(&now, NULL);
@@ -100,8 +100,8 @@ int test(void)
 			}
 		}
 		syslog(LOG_DEBUG, "get_sendto_fds() returned %d", i);
-		syslog(LOG_DEBUG, "select(%d, NULL, xx, NULL, %ld.%06ld)",
-		       max_fd, (long)timeout.tv_sec, (long)timeout.tv_usec);
+		syslog(LOG_DEBUG, "select(%d, NULL, xx, NULL, %lld.%06ld)",
+		       max_fd, (long long)timeout.tv_sec, (long)timeout.tv_usec);
 		i = select(max_fd, NULL, &writefds, NULL, &timeout);
 		if(i < 0) {
 			syslog(LOG_ERR, "select: %m");
